requirements.txt
bcrypt==4.2.0
blinker~=1.6.3
click==8.1.7
Flask==3.0.3
itsdangerous~=2.1.2
Jinja2~=3.1.2
MarkupSafe~=2.1.3
mysql-connector-python~=8.1.0
Werkzeug~=3.0.1

pip~=24.0
attrs~=23.1.0
DateTime~=5.2
logging~=0.4.9.6
distro~=1.8.0
wheel~=0.43.0
protobuf~=4.21.12
pillow~=10.2.0
tornado~=6.4
filelock~=3.15.4
sniffio~=1.3.0
typing_extensions~=4.8.0
dnspython~=2.6.1
mysql-connector~=2.2.9
sqlparse~=0.5.1
fastapi~=0.111.1
starlette~=0.37.2
email_validator~=2.2.0
pydantic~=2.5.0
pydantic_core~=2.14.1
requests~=2.32.3
websockets~=12.0
anyio~=3.7.1
ujson~=5.10.0
orjson~=3.10.6
Deprecated~=1.2.14
annotated-types~=0.6.0
PyYAML~=6.0.1
python-multipart~=0.0.9
httpx~=0.25.2